Fixed color to float

+++ classPlanet.cpp
@@ -13,7 +13,7 @@ private:
     array<double, 2> position; // {x, y}
     array<double, 2> velocity; // {vx, vy}
     array<double, 2> acceleration; // {ax, ay}
-    array<int, 4> color // {R, G, B, A}
+    array<float, 4> color // {R, G, B, A}
 
 public:
   Planet(string name, double mass, double radius, array<double, 2> position, array<double, 2> velocity, array<double,2> acceleration, array<int, 4> color)
@@ -26,7 +26,7 @@ double getRadius() const { return (radius); }
 array<double, 2> getPosition() const { return position; }
 array<double, 2> getVelocity() const { return velocity; }
 array<double, 2> getAcceleration() const { return acceleration; }
-array<int, 4> getColor() const { return color; }
+array<float, 4> getColor() const { return color; }
 
 //Setters
 void setName(string name) { this->name = name; }
@@ -35,7 +35,7 @@ void setRadius(double radius) { this->radius = radius; }
 void setPosition(array<double, 2> position) { this->position = position; }
 void setVelocity(array<double, 2> velocity) { this->velocity = velocity; }
 void setAcceleration(array<double, 2> acceleration) { this->acceleration = acceleration; }
-void setColor(array<int, 4> color) { this->color = color; }
+void setColor(array<float, 4> color) { this->color = color; }
 };
